fix small truncation warning in IPv4Endpoint
#rb Jason.Walter
#jira UE-166274
#preflight 63c06647a06ab115ea7aeec9

[CL 23668346 by Francis Hurteau in ue5-main branch]

+++ Engine/Source/Runtime/Networking/Public/Interfaces/IPv4/IPv4Endpoint.h
@@ -61,7 +61,7 @@ public:
 		int32 OutPort;
 		InternetAddr->GetIp(Address.Value);
 		InternetAddr->GetPort(OutPort);
-		Port = OutPort;
+		Port = static_cast<uint16>(OutPort);
 	}
 
 public:
